// frontend/src/api/generated/models/projectStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus]

export const ProjectStatus = {
  Planning: 'Planning',
  In_Progress: 'In Progress',
  On_Hold: 'On Hold',
  Completed: 'Completed',
} as const